Remove Gherkin namespace
